src/services/schema_parser.py

import json
import yaml
import csv
from pathlib import Path
from typing import Dict, List, Any, Union


class SchemaParser:
    """Parser for schema files (JSON, YAML, CSV) to normalized JSON format."""

    def __init__(self):
        pass
    self.supported_formats = ['.json', '.yaml', '.yml', '.csv']

    def parse_file(self, input_path: Union[str, Path]) -> Dict[str, Any]:
        pass
    """Parse a schema file and return normalized JSON data.

 Args:
 input_path: Path to the schema file

 Returns:
 Dict containing parsed schema data

 Raises:
 FileNotFoundError: If the input file doesn't exist
 ValueError: If the file format is not supported
 """
    input_path = Path(input_path)

    if not input_path.exists():
        pass
    raise FileNotFoundError(f"Schema file not found at: {input_path}")

    if input_path.suffix not in self.supported_formats:
        pass
    raise ValueError(
        f"Unsupported file format {
            input_path.suffix}. Supported formats: {
            self.supported_formats}")

    # Parse based on file extension
    if input_path.suffix == ".json":
        pass
    return self._parse_json(input_path)
    elif input_path.suffix in [".yaml", ".yml"]:
        pass
    return self._parse_yaml(input_path)
    elif input_path.suffix == ".csv":
        pass
    return self._parse_csv(input_path)

    def _parse_json(self, file_path: Path) -> Dict[str, Any]:
        pass
    """Parse JSON file."""
    with open(file_path, 'r') as f:
        pass
    return json.load(f)

    def _parse_yaml(self, file_path: Path) -> Dict[str, Any]:
        pass
    """Parse YAML file."""
    with open(file_path, 'r') as f:
        pass
    return yaml.safe_load(f)

    def _parse_csv(self, file_path: Path) -> List[Dict[str, Any]]:
        pass
    """Parse CSV file."""
    with open(file_path, 'r') as f:
        pass
    reader = csv.DictReader(f)
    return [row for row in reader]

    def parse_to_file(
            self, input_path: Union[str, Path], output_path: Union[str, Path]) -> None:
    """Parse a schema file and save to JSON file.

 Args:
 input_path: Path to the input schema file
 output_path: Path to save the normalized JSON file
 """
    schema_data = self.parse_file(input_path)
    output_path = Path(output_path)

    with open(output_path, 'w') as f:
        pass
    json.dump(schema_data, f, indent=4)

    print(f"[PASS] Schema successfully parsed and saved to {output_path}")

    def validate_schema(self, schema_data: Dict[str, Any]) -> bool:
        pass
    """Validate that schema contains required fields.

 Args:
 schema_data: Parsed schema data

 Returns:
 True if schema is valid, False otherwise
 """
    required_fields = ['class_id', 'expected_topics']

    if isinstance(schema_data, list):
        pass
        # CSV format returns list of dicts
    if not schema_data:
        pass
    return False
    return all(field in schema_data[0] for field in required_fields)
    elif isinstance(schema_data, dict):
        pass
    return all(field in schema_data for field in required_fields)

    return False


def parse_schema_to_json(input_path: Path, output_path: Path):
    """
    Parses a schema file (JSON, YAML, or CSV) and converts it to a normalized JSON format.

    Args:
    input_path (Path): The path to the input schema file.
    output_path (Path): The path to save the normalized JSON file.
    """

    if not input_path.exists():
        pass
    raise FileNotFoundError(f"Schema file not found at: {input_path}")

    # Read and parse based on file extension
    if input_path.suffix == ".json":
        pass
    with open(input_path, 'r') as f:
        pass
    schema_data = json.load(f)
    elif input_path.suffix in [".yaml", ".yml"]:
        pass
    with open(input_path, 'r') as f:
        pass
    schema_data = yaml.safe_load(f)
    elif input_path.suffix == ".csv":
        pass
    with open(input_path, 'r') as f:
        pass
    reader = csv.DictReader(f)
    schema_data = [row for row in reader]
    else:
        pass
    raise ValueError("Unsupported file format. Please use JSON, YAML, or CSV.")

    # Save to normalized JSON
    with open(output_path, 'w') as f:
        pass
    json.dump(schema_data, f, indent=4)

    print(f"[PASS] Schema successfully parsed and saved to {output_path}")


if __name__ == '__main__':
    pass
    # Example usage (for testing)
    # Create dummy schema files for testing
    dummy_json_path = Path("dummy_schema.json")
    dummy_yaml_path = Path("dummy_schema.yaml")
    dummy_csv_path = Path("dummy_schema.csv")

    with open(dummy_json_path, 'w') as f:
        pass
    json.dump({"class_id": "CS101", "date": "2023-10-27",
              "expected_topics": ["Intro", "Variables"]}, f)

    with open(dummy_yaml_path, 'w') as f:
        pass
    yaml.dump({"class_id": "CS101", "date": "2023-10-27",
              "expected_topics": ["Intro", "Variables"]}, f)

    with open(dummy_csv_path, 'w') as f:
        pass
    writer = csv.writer(f)
    writer.writerow(["class_id", "date", "expected_topics"])
    writer.writerow(["CS101", "2023-10-27", "Intro,Variables"])

    # Test the parser
    try:
        pass
    parse_schema_to_json(dummy_json_path, Path("expected_topics.json"))
    parse_schema_to_json(
        dummy_yaml_path,
        Path("expected_topics_from_yaml.json"))
    parse_schema_to_json(dummy_csv_path, Path("expected_topics_from_csv.json"))
    finally:
        pass
        # Clean up dummy files
    dummy_json_path.unlink()
    dummy_yaml_path.unlink()
    dummy_csv_path.unlink()
