requirements.txt
# =====================================================
# Live Camera Enhanced Translator - Requirements
# =====================================================

# Core Framework and UI
gradio>=5.35.0
fastapi>=0.115.0
uvicorn>=0.35.0
starlette>=0.46.0
flask>=3.0.0
flask-cors>=4.0.0

# Audio/Video Processing
openai-whisper>=20250625
faster-whisper>=1.1.0
ctranslate2>=4.0.0
onnxruntime>=1.20.0
pyaudio>=0.2.14
av>=14.4.0
opencv-contrib-python>=4.11.0
sounddevice>=0.5.2
soundfile>=0.13.1
pydub>=0.25.1
librosa>=0.10.2
webrtcvad>=2.0.10

# AI/ML Dependencies
google-generativeai>=0.8.5
torch>=2.7.0
torchaudio>=2.7.0
transformers>=4.53.0
numpy>=2.2.0
scipy>=1.16.0
scikit-learn>=1.5.0
accelerate>=1.8.0

# Data Processing
pandas>=2.3.0
PyYAML>=6.0.0
orjson>=3.10.0
python-dateutil>=2.9.0

# File Processing
pillow>=11.3.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
markdown>=3.5.0

# Networking and Communication
websockets>=15.0.1
websocket-client>=1.8.0
aiohttp>=3.12.0
httpx>=0.28.0
requests>=2.32.0

# Redis and Caching
redis>=6.2.0
hiredis>=2.0.0

# NLP and Language Processing
spacy>=3.8.0
nltk>=3.9.0
rake-nltk>=1.0.6
tiktoken>=0.9.0
tokenizers>=0.21.0

# Development and Testing
pytest>=8.4.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
black>=24.10.0
flake8>=7.1.0
isort>=5.13.0
mypy>=1.13.0

# Utilities
tqdm>=4.67.0
click>=8.2.0
typer>=0.16.0
rich>=14.0.0
python-dotenv>=1.0.0
filelock>=3.18.0
fsspec>=2025.5.0
packaging>=25.0.0

# Security
cryptography>=45.0.0
pyOpenSSL>=25.1.0
certifi>=2025.6.0

# Optional: GPU Support (uncomment if using CUDA)
# torch>=2.7.0+cu118
# torchaudio>=2.7.0+cu118

# Optional: Advanced Features (uncomment if needed)
# tensorboard>=2.19.0
# prometheus-client>=0.21.0
# sentry-sdk>=2.21.0

# Python Version Requirement
# Requires Python 3.8+
